[Opt Plus Only Feature] Allow users to disable join flipping
There are cases where users want to disable cost-based join
re-ordering, but still want the join type to be automatically
determined. Before this change, it was possible for the
DetermineJoinDistributionType rule to flip the join order
using statistics when it was undesirable.

This adds a new feature config for the coordinator and a
corresponding session property.

- feature config: optimizer.size-based-join-flipping-enabled
- session property: optimizer_size_based_join_flipping_enabled

By default, this is true to preserve the previous behavior, but setting
it to false will disable returning any results which flip the sides of
the join in the DetermineJoinDistributionType rule.

Cherry-pick of bcaa3f1

Alchemy-item: [[Opt Plus Only Feature] Allow users to disable join flipping](https://github.ibm.com/lakehouse/presto/pull/3382#issuecomment-118074759) commit 1/1 - 4429263957dafa084b4b9f968a2dd9feaf5e385a