[native] Fix arrow-flight to be used for IBM internal use
In OSS the name is not branded but internally we use ibm-arrow-flight
instead of arrow-flight because on the Java side the arrow flight
plugin is also called ibm-arrow-flight and keeping a consistent
name in the catalog configs for the connector helps simplify the
configuration.

We also need to link into the custom IBM authenticators.

Alchemy-item: [[native] Fix arrow-flight to be used for IBM internal use](https://github.ibm.com/lakehouse/presto/pull/3449#issuecomment-117431248) commit 1/1 - a70266ec0bde9a7a8409efb42abaa75dedac998e